fix: AccessToken 갱신 API 수정 (#60) (KAN-111)
