+++ src/main/java/org/ioteatime/meonghanyangserver/auth/controller/AuthApi.java
@@ -4,14 +4,12 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.ioteatime.meonghanyangserver.auth.dto.reponse.LoginResponse;
-import org.ioteatime.meonghanyangserver.auth.dto.reponse.RefreshResponse;
 import org.ioteatime.meonghanyangserver.auth.dto.request.EmailRequest;
 import org.ioteatime.meonghanyangserver.auth.dto.request.LoginRequest;
 import org.ioteatime.meonghanyangserver.auth.dto.request.VerifyEmailRequest;
 import org.ioteatime.meonghanyangserver.common.api.Api;
 import org.ioteatime.meonghanyangserver.user.dto.request.JoinRequest;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 
 @Tag(name = "Auth Api", description = "인증 관련 API 목록입니다.")
 public interface AuthApi {
@@ -29,7 +27,4 @@ public interface AuthApi {
 
     @Operation(summary = "이메일 중복을 확인 합니다.")
     Api<?> duplicateEmail(@Valid @RequestBody EmailRequest email);
-
-    @Operation(summary = "토큰을 다시 생성합니다.")
-    Api<RefreshResponse> refreshToken(@RequestHeader("Authorization") String authorizationHeader);
 }
