+++ src/main/java/org/ioteatime/meonghanyangserver/auth/dto/reponse/RefreshResponse.java
@@ -1,6 +1,9 @@
 package org.ioteatime.meonghanyangserver.auth.dto.reponse;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
-import lombok.*;
 
-public record RefreshResponse(@NotNull String accessToken) {}
+@Schema(description = "갱신된 AccessToken 입니다.")
+public record RefreshResponse(
+        @NotNull @Schema(description = "갱신된 AccessToken", example = "Bearer wefewfwe...")
+                String accessToken) {}
