+++ src/main/java/org/ioteatime/meonghanyangserver/common/type/AuthErrorType.java
@@ -9,7 +9,8 @@ public enum AuthErrorType implements ErrorTypeCode {
     HEADER_INVALID("UNAUTHORIZED", "AUTHORIZATION 헤더가 누락되었거나 토큰 형식에 오류가 있습니다."),
     PASSWORD_INVALID("BAD REQUEST", "비밀번호가 유효하지 않습니다."),
     CODE_NOT_EQUALS("BAD REQUEST", "이메일 인증 코드가 일치하지 않습니다."),
-    EMAIL_DUPLICATED("BAD REQUEST", "이메일이 중복되었습니다.");
+    EMAIL_DUPLICATED("BAD REQUEST", "이메일이 중복되었습니다."),
+    REFRESH_TOKEN_NOT_FOUND("NOT FOUND", "REFRESH TOKEN을 찾을 수 없습니다.");
 
     private final String message;
     private final String description;
