+++ src/main/java/org/ioteatime/meonghanyangserver/user/service/UserService.java
@@ -1,9 +1,14 @@
 package org.ioteatime.meonghanyangserver.user.service;
 
 import lombok.RequiredArgsConstructor;
+import org.ioteatime.meonghanyangserver.auth.dto.reponse.RefreshResponse;
+import org.ioteatime.meonghanyangserver.auth.mapper.AuthResponseMapper;
 import org.ioteatime.meonghanyangserver.common.exception.BadRequestException;
 import org.ioteatime.meonghanyangserver.common.exception.NotFoundException;
 import org.ioteatime.meonghanyangserver.common.type.AuthErrorType;
+import org.ioteatime.meonghanyangserver.common.utils.JwtUtils;
+import org.ioteatime.meonghanyangserver.redis.RefreshToken;
+import org.ioteatime.meonghanyangserver.redis.RefreshTokenRepository;
 import org.ioteatime.meonghanyangserver.user.domain.UserEntity;
 import org.ioteatime.meonghanyangserver.user.dto.request.ChangePasswordRequest;
 import org.ioteatime.meonghanyangserver.user.dto.response.UserDetailResponse;
@@ -16,9 +21,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class UserService {
-
+    private final JwtUtils jwtUtils;
     private final UserRepository userRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
+    private final RefreshTokenRepository refreshTokenRepository;
 
     public UserDetailResponse getUserDetail(Long userId) {
         UserEntity userEntity =
@@ -49,5 +55,34 @@ public class UserService {
 
         // Dirty-Checking Password Change
         userEntity.setPassword(bCryptPasswordEncoder.encode(newPassword));
+    }
+
+    public RefreshResponse reissueAccessToken(String authorizationHeader) {
+        String refreshToken = jwtUtils.extractTokenFromHeader(authorizationHeader);
+
+        Long userId = jwtUtils.getIdFromToken(refreshToken);
+        UserEntity userEntity =
+                userRepository
+                        .findById(userId)
+                        .orElseThrow(() -> new NotFoundException(AuthErrorType.NOT_FOUND));
+
+        if (!jwtUtils.validateToken(refreshToken, userEntity)) {
+            throw new NotFoundException(AuthErrorType.REFRESH_TOKEN_INVALID);
+        }
+
+        RefreshToken storedToken =
+                refreshTokenRepository
+                        .findByRefreshToken(refreshToken)
+                        .orElseThrow(
+                                () -> new NotFoundException(AuthErrorType.REFRESH_TOKEN_NOT_FOUND));
+
+        if (!storedToken.getRefreshToken().equals(refreshToken)) {
+            throw new BadRequestException(AuthErrorType.TOKEN_NOT_EQUALS);
+        }
+
+        String newAccessToken = jwtUtils.generateAccessToken(userEntity);
+        newAccessToken = jwtUtils.includeBearer(newAccessToken);
+
+        return AuthResponseMapper.from(newAccessToken);
     }
 }
